pyproject.toml
[tool.poetry]
name = "polus-plugins-data-rxiv"
version = "0.1.0"
description = ""
authors = ["nicholas-schaub <nicholas.j.schaub@gmail.com>"]
readme = "README.md"
packages = [{include = "polus", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "1.10.4"
requests = "^2.31.0"
rxiv-types = "^0.1.0"
boto3 = "^1.33.11"
python-dotenv = "^1.0.0"
torch = "^2.1.1"
torchvision = "^0.16.1"
torchaudio = "^2.1.1"
instructorembedding = "^1.0.1"
faiss-gpu = "^1.7.2"
langchain = "^0.0.350"
sentence-transformers = "^2.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
mypy = "^1.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
